requirements.txt
fastapi
uvicorn[standard]
psutil
watchdog
scikit-learn
numpy
sqlalchemy
aiosqlite
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
pyyaml
jinja2
aiofiles
websockets
rich
httpx
pytest
pytest-asyncio